changes to number

+++ src/app/modules/purchase-items/entities/item-stock-track.entity.ts
@@ -38,10 +38,10 @@ export class ItemsStockTrack {
   @Column({ type: "varchar", nullable: true })
   stockNumber: string;
 
-  @Column({ type: "int", nullable: false })
+  @Column({ type: "decimal", nullable: false })
   quantityAdded: number;
 
-  @Column({ type: "int", nullable: false })
+  @Column({ type: "decimal", nullable: false })
   quantityUvailable: number;
 
   @CreateDateColumn({ type: "varchar", nullable: false })
